Initial Creation Of Project
In this instance, I created the Final Project folder and completely got Express downloaded and enabled. However, with tailwindcss I got it downloaded as well although for some odd reason it won't initialize properly. I've spent the last 30 to 40 minutes trying to find a solution to no avail, so tailwindcss is only partially implemented at the time being.